Add unit tests for various algorithms in the data structure and skills categories
- Implement tests for tree serialization and deserialization.
- Add tests for converting sorted arrays to binary search trees.
- Create tests for word search using backtracking.
- Include tests for generating parentheses and letter combinations.
- Add tests for permutations and subsets using backtracking.
- Implement tests for solving Sudoku puzzles.
- Create tests for counting ways to group overlapping ranges.
- Add tests for merging intervals and merging sorted arrays.
- Implement tests for designing a graph with shortest path calculations.
- Add tests for binary search algorithms including finding peaks and bad versions.
- Create tests for searching in a 2D matrix and finding ranges.
- Add tests for bit manipulation functions including sum, Hamming distance, and missing numbers.
- Implement tests for string manipulation functions including reversing strings and finding palindromes.
- Add tests for two-pointer techniques including moving zeros and removing duplicates.
- Create tests for three and four sum problems.

+++ workspace/src/data_structure/array_and_string/maximize_sum.cpp
@@ -31,7 +31,7 @@ public:
     int maximizeSum(std::vector<int>& nums, int k) {
 
         int max = nums[0];
-        for (int i = 1; i < nums.size(); i++) {
+        for (int i = 1; i < (int)nums.size(); i++) {
             if (max < nums[i])
                 max = nums[i];
         }
